Find alternative products for all products containing certain active substances

This function searches for uses of a given list of active substances and reports either a table of uses with the number of available alternative products for each use, a detailed table of the alternative product uses, a table of uses without alternatives, or a list containing these three tables.

Usage

alternative_products(
  srppp,
  active_ingredients,
  details = FALSE,
  missing = FALSE,
  list = FALSE,
  lang = c("de", "fr", "it")
)

Arguments

srppp

A srppp_dm object.

active_ingredients

Character vector of active ingredient names that will be matched against the column 'substances_de' in the srppp table 'substances'.

details

Should a table of alternative uses with 'wNbr' and 'use_nr' be returned?

missing

If this is set to TRUE, uses without alternative product registrations are listed.

list

If TRUE, a list of three tables is returned, a table of uses without alternative products ("Lückenindikationen"), a table of the number of alternative products for each use, if any, and a detailed table of all the alternative uses. This argument overrides the arguments 'details' and 'missing'.

lang

The language used for the active ingredient names and the returned tables.

Value

A tibble::tibble containing use definitions as defined above, i.e. containing columns with the application area, crop and pathogen. Depending on the arguments, columns summarizing or listing the alternative products and/or uses are also contained.

Details

A use is defined here as a combination of an application area, a crop ('culture') and a pathogen ('pest'). This means, that for an alternative product to be found, there has to be an exact match of application area, crop an pathogen.

Examples

# \donttest{
sr <- srppp_dm()

actives_de <- c("Lambda-Cyhalothrin", "Deltamethrin")

alternative_products(sr, actives_de)
#> # A tibble: 446 × 5
#>    application_area_de culture_de              pest_de             n_wNbr n_pNbr
#>    <chr>               <chr>                   <chr>                <int>  <int>
#>  1 Beerenbau           Erdbeere                Erdbeer- oder Himb…     16      7
#>  2 Beerenbau           Erdbeere                Thripse                 31     16
#>  3 Beerenbau           Himbeere                Erdbeer- oder Himb…     11      3
#>  4 Beerenbau           Himbeere                Himbeerkäfer            11      3
#>  5 Feldbau             Ackerbohne              Erdraupen                5      4
#>  6 Feldbau             Eiweisserbse            Erbsenblattrandkäf…      0      0
#>  7 Feldbau             Eiweisserbse            Erbsenwickler            4      1
#>  8 Feldbau             Eiweisserbse            Erdraupen                5      4
#>  9 Feldbau             Futter- und Zuckerrüben Blattläuse (Röhren…      6      2
#> 10 Feldbau             Futter- und Zuckerrüben Erdraupen                5      4
#> # ℹ 436 more rows
alternative_products(sr, actives_de, missing = TRUE)
#> # A tibble: 108 × 3
#>    application_area_de culture_de                       pest_de                 
#>    <chr>               <chr>                            <chr>                   
#>  1 Feldbau             Eiweisserbse                     Erbsenblattrandkäfer    
#>  2 Feldbau             Futter- und Zuckerrüben          Rübenfliege             
#>  3 Feldbau             Futterrübe                       Erdraupen               
#>  4 Feldbau             Futterrübe                       Rübenerdflöhe           
#>  5 Feldbau             Getreide                         Virusübertragende Blatt…
#>  6 Feldbau             Lagerhallen, Mühlen, Silogebäude Vorratsschädlinge       
#>  7 Feldbau             Mais                             Fritfliege              
#>  8 Feldbau             Sojabohne                        Distelfalter            
#>  9 Feldbau             Wintergetreide                   Virusübertragende Blatt…
#> 10 Feldbau             Zuckerrübe                       Erdraupen               
#> # ℹ 98 more rows
alternative_products(sr, actives_de, details = TRUE)
#> # A tibble: 5,798 × 6
#>    application_area_de culture_de pest_de                      pNbr wNbr  use_nr
#>    <chr>               <chr>      <chr>                       <int> <chr>  <int>
#>  1 Beerenbau           Erdbeere   Erdbeer- oder Himbeerblüte…  4426 4343      17
#>  2 Beerenbau           Erdbeere   Erdbeer- oder Himbeerblüte…  4568 4491      12
#>  3 Beerenbau           Erdbeere   Erdbeer- oder Himbeerblüte…  4568 4491…     12
#>  4 Beerenbau           Erdbeere   Erdbeer- oder Himbeerblüte…  7036 6020      37
#>  5 Beerenbau           Erdbeere   Erdbeer- oder Himbeerblüte…  7036 6020…     37
#>  6 Beerenbau           Erdbeere   Erdbeer- oder Himbeerblüte…  7036 6020…     37
#>  7 Beerenbau           Erdbeere   Erdbeer- oder Himbeerblüte…  8464 7133       7
#>  8 Beerenbau           Erdbeere   Erdbeer- oder Himbeerblüte…  8464 7133…      7
#>  9 Beerenbau           Erdbeere   Erdbeer- oder Himbeerblüte…  8464 7133…      7
#> 10 Beerenbau           Erdbeere   Erdbeer- oder Himbeerblüte…  8800 7106       8
#> # ℹ 5,788 more rows
alternative_products(sr, actives_de, list = TRUE)
#> $`No alternative`
#> # A tibble: 108 × 3
#>    application_area_de culture_de                       pest_de                 
#>    <chr>               <chr>                            <chr>                   
#>  1 Feldbau             Eiweisserbse                     Erbsenblattrandkäfer    
#>  2 Feldbau             Futter- und Zuckerrüben          Rübenfliege             
#>  3 Feldbau             Futterrübe                       Erdraupen               
#>  4 Feldbau             Futterrübe                       Rübenerdflöhe           
#>  5 Feldbau             Getreide                         Virusübertragende Blatt…
#>  6 Feldbau             Lagerhallen, Mühlen, Silogebäude Vorratsschädlinge       
#>  7 Feldbau             Mais                             Fritfliege              
#>  8 Feldbau             Sojabohne                        Distelfalter            
#>  9 Feldbau             Wintergetreide                   Virusübertragende Blatt…
#> 10 Feldbau             Zuckerrübe                       Erdraupen               
#> # ℹ 98 more rows
#> 
#> $`Number of alternatives`
#> # A tibble: 446 × 5
#>    application_area_de culture_de              pest_de             n_wNbr n_pNbr
#>    <chr>               <chr>                   <chr>                <int>  <int>
#>  1 Beerenbau           Erdbeere                Erdbeer- oder Himb…     16      7
#>  2 Beerenbau           Erdbeere                Thripse                 31     16
#>  3 Beerenbau           Himbeere                Erdbeer- oder Himb…     11      3
#>  4 Beerenbau           Himbeere                Himbeerkäfer            11      3
#>  5 Feldbau             Ackerbohne              Erdraupen                5      4
#>  6 Feldbau             Eiweisserbse            Erbsenblattrandkäf…      0      0
#>  7 Feldbau             Eiweisserbse            Erbsenwickler            4      1
#>  8 Feldbau             Eiweisserbse            Erdraupen                5      4
#>  9 Feldbau             Futter- und Zuckerrüben Blattläuse (Röhren…      6      2
#> 10 Feldbau             Futter- und Zuckerrüben Erdraupen                5      4
#> # ℹ 436 more rows
#> 
#> $`Alternative uses`
#> # A tibble: 5,798 × 6
#>    application_area_de culture_de pest_de                      pNbr wNbr  use_nr
#>    <chr>               <chr>      <chr>                       <int> <chr>  <int>
#>  1 Beerenbau           Erdbeere   Erdbeer- oder Himbeerblüte…  4426 4343      17
#>  2 Beerenbau           Erdbeere   Erdbeer- oder Himbeerblüte…  4568 4491      12
#>  3 Beerenbau           Erdbeere   Erdbeer- oder Himbeerblüte…  4568 4491…     12
#>  4 Beerenbau           Erdbeere   Erdbeer- oder Himbeerblüte…  7036 6020      37
#>  5 Beerenbau           Erdbeere   Erdbeer- oder Himbeerblüte…  7036 6020…     37
#>  6 Beerenbau           Erdbeere   Erdbeer- oder Himbeerblüte…  7036 6020…     37
#>  7 Beerenbau           Erdbeere   Erdbeer- oder Himbeerblüte…  8464 7133       7
#>  8 Beerenbau           Erdbeere   Erdbeer- oder Himbeerblüte…  8464 7133…      7
#>  9 Beerenbau           Erdbeere   Erdbeer- oder Himbeerblüte…  8464 7133…      7
#> 10 Beerenbau           Erdbeere   Erdbeer- oder Himbeerblüte…  8800 7106       8
#> # ℹ 5,788 more rows
#> 

# Example in Italian
actives_it <- c("Lambda-Cialotrina", "Deltametrina")
alternative_products(sr, actives_it, lang = "it")
#> # A tibble: 446 × 5
#>    application_area_it culture_it                          pest_it n_wNbr n_pNbr
#>    <chr>               <chr>                               <chr>    <int>  <int>
#>  1 Campicoltura        Barbabietola da foraggio            Altich…      0      0
#>  2 Campicoltura        Barbabietola da foraggio            Nottue…      0      0
#>  3 Campicoltura        Barbabietola da zucchero            Altich…      0      0
#>  4 Campicoltura        Barbabietola da zucchero            Nottue…      0      0
#>  5 Campicoltura        Barbabietole da foraggio e da zucc… Afidi        6      2
#>  6 Campicoltura        Barbabietole da foraggio e da zucc… Altich…      5      4
#>  7 Campicoltura        Barbabietole da foraggio e da zucc… Mosca …      0      0
#>  8 Campicoltura        Barbabietole da foraggio e da zucc… Nottue…      5      4
#>  9 Campicoltura        Barbabietole da foraggio e da zucc… tignol…      2      2
#> 10 Campicoltura        Cartamo                             Nottue…      5      4
#> # ℹ 436 more rows
# }