Resolve Cultures Function

Resolves culture levels in a dataset to their lowest hierarchical level (leaf nodes) using a parent-child relationship dataset derived from a culture tree. If no match is found, the function assigns NA to the leaf_culture_de column.

Usage

resolve_cultures(
  dataset,
  srppp,
  culture_column = "culture_de",
  correct_culture_names = TRUE
)

Arguments

dataset

A data frame or tibble containing the data to be processed. It should include a column that represents the culture information to be resolved.

srppp

An srppp_dm object. From this object the parent child relations from the culture tree are used, which are stored as attribute 'parent_child_df' of the culture tree, which is itself stored as an attribute of the object.

culture_column

(Optional) A character string specifying the column in the dataset that contains the culture information to be resolved. Defaults to "culture_de".

correct_culture_names

If this argument is set to TRUE, the following corrections will be applied: In the culture_tree, and consequently in the parent_child_df, there are variations in the naming of aggregated culture groups with "allg.". For example, both "Obstbau allg." and "allg. Obstbau" exist. However, information about the leaf nodes is only available in the culture groups that start with "allg. ...". This will be adjusted.

Value

A data frame or tibble with the same structure as the input dataset, but with an additional column "leaf_culture_de" that contains the resolved leaf culture levels.

Details

The resolve_cultures function processes the input dataset as follows

Leaf Node Resolution: The cultures in the specified column of the dataset are resolved to their lowest hierarchical level (leaf nodes) based on the parent_child_df mapping. For each row in the dataset, the function searches for a match in parent_child_df: - If a match is found, new rows are generated for each corresponding leaf culture. - If no match is found, the leaf_culture_de column is set to NA.

The result is an expanded dataset that includes an additional column (leaf_culture_de) containing the resolved cultures at their lowest level.

Examples

# \donttest{
example_dataset_1 <- data.frame(
  substance_de = c("Spirotetramat", "Spirotetramat", "Spirotetramat", "Spirotetramat"),
  pNbr = c(7839, 7839, 7839, 7839),
  use_nr = c(5, 7, 18, 22),
  application_area_de = c("Obstbau", "Obstbau", "Obstbau", "Obstbau"),
  culture_de = c("Birne", "Kirsche", "Steinobst", "Kernobst"),
    pest_de = c("Birnblattsauger", "Kirschenfliege", "Blattläuse (Röhrenläuse)", "Spinnmilben")
    )

example_dataset_2 <- data.frame(
  substance_de = c("Spirotetramat", "Spirotetramat", "Spirotetramat", "Spirotetramat"),
  pNbr = c(7839, 7839, 7839, 7839),
  use_nr = c(5, 7, 18, 22),
  application_area_de = c("Obstbau", "Obstbau", "Obstbau", "Obstbau"),
  culture_de = c("Birne", "Kirschen", "Steinobst", "Obstbau allg."),
    pest_de = c("Birnblattsauger", "Kirschenfliege", "Blattläuse (Röhrenläuse)", "Spinnmilben")
    )

library(srppp)
current_register <- srppp_dm()

result1 <- resolve_cultures(example_dataset_1, current_register)
print(result1)
#>     substance_de pNbr use_nr application_area_de culture_de
#> 1  Spirotetramat 7839      5             Obstbau      Birne
#> 2  Spirotetramat 7839      7             Obstbau    Kirsche
#> 3  Spirotetramat 7839     18             Obstbau  Steinobst
#> 31 Spirotetramat 7839     18             Obstbau  Steinobst
#> 32 Spirotetramat 7839     18             Obstbau  Steinobst
#> 33 Spirotetramat 7839     18             Obstbau  Steinobst
#> 34 Spirotetramat 7839     18             Obstbau  Steinobst
#> 4  Spirotetramat 7839     22             Obstbau   Kernobst
#> 41 Spirotetramat 7839     22             Obstbau   Kernobst
#> 42 Spirotetramat 7839     22             Obstbau   Kernobst
#>                     pest_de      leaf_culture_de
#> 1           Birnblattsauger                Birne
#> 2            Kirschenfliege              Kirsche
#> 3  Blattläuse (Röhrenläuse)             Aprikose
#> 31 Blattläuse (Röhrenläuse)              Kirsche
#> 32 Blattläuse (Röhrenläuse) Pfirsich / Nektarine
#> 33 Blattläuse (Röhrenläuse)              Pflaume
#> 34 Blattläuse (Röhrenläuse)            Zwetschge
#> 4               Spinnmilben                Apfel
#> 41              Spinnmilben               Quitte
#> 42              Spinnmilben                Birne
result2 <- resolve_cultures(example_dataset_2, current_register)
print(result2)
#>     substance_de pNbr use_nr application_area_de    culture_de
#> 1  Spirotetramat 7839      5             Obstbau         Birne
#> 2  Spirotetramat 7839      7             Obstbau      Kirschen
#> 3  Spirotetramat 7839     18             Obstbau     Steinobst
#> 31 Spirotetramat 7839     18             Obstbau     Steinobst
#> 32 Spirotetramat 7839     18             Obstbau     Steinobst
#> 33 Spirotetramat 7839     18             Obstbau     Steinobst
#> 34 Spirotetramat 7839     18             Obstbau     Steinobst
#> 4  Spirotetramat 7839     22             Obstbau Obstbau allg.
#> 41 Spirotetramat 7839     22             Obstbau Obstbau allg.
#> 42 Spirotetramat 7839     22             Obstbau Obstbau allg.
#> 43 Spirotetramat 7839     22             Obstbau Obstbau allg.
#> 44 Spirotetramat 7839     22             Obstbau Obstbau allg.
#> 45 Spirotetramat 7839     22             Obstbau Obstbau allg.
#> 46 Spirotetramat 7839     22             Obstbau Obstbau allg.
#> 47 Spirotetramat 7839     22             Obstbau Obstbau allg.
#> 48 Spirotetramat 7839     22             Obstbau Obstbau allg.
#> 49 Spirotetramat 7839     22             Obstbau Obstbau allg.
#>                     pest_de culture_de_corrected      leaf_culture_de
#> 1           Birnblattsauger                Birne                Birne
#> 2            Kirschenfliege             Kirschen                 <NA>
#> 3  Blattläuse (Röhrenläuse)            Steinobst             Aprikose
#> 31 Blattläuse (Röhrenläuse)            Steinobst              Kirsche
#> 32 Blattläuse (Röhrenläuse)            Steinobst Pfirsich / Nektarine
#> 33 Blattläuse (Röhrenläuse)            Steinobst              Pflaume
#> 34 Blattläuse (Röhrenläuse)            Steinobst            Zwetschge
#> 4               Spinnmilben        allg. Obstbau                Olive
#> 41              Spinnmilben        allg. Obstbau                Apfel
#> 42              Spinnmilben        allg. Obstbau               Quitte
#> 43              Spinnmilben        allg. Obstbau             Aprikose
#> 44              Spinnmilben        allg. Obstbau              Kirsche
#> 45              Spinnmilben        allg. Obstbau Pfirsich / Nektarine
#> 46              Spinnmilben        allg. Obstbau                Birne
#> 47              Spinnmilben        allg. Obstbau              Pflaume
#> 48              Spinnmilben        allg. Obstbau            Zwetschge
#> 49              Spinnmilben        allg. Obstbau              Walnuss
result3 <- resolve_cultures(example_dataset_2, current_register,
  correct_culture_names = FALSE)
print(result3)
#>     substance_de pNbr use_nr application_area_de    culture_de
#> 1  Spirotetramat 7839      5             Obstbau         Birne
#> 2  Spirotetramat 7839      7             Obstbau      Kirschen
#> 3  Spirotetramat 7839     18             Obstbau     Steinobst
#> 31 Spirotetramat 7839     18             Obstbau     Steinobst
#> 32 Spirotetramat 7839     18             Obstbau     Steinobst
#> 33 Spirotetramat 7839     18             Obstbau     Steinobst
#> 34 Spirotetramat 7839     18             Obstbau     Steinobst
#> 4  Spirotetramat 7839     22             Obstbau Obstbau allg.
#>                     pest_de      leaf_culture_de
#> 1           Birnblattsauger                Birne
#> 2            Kirschenfliege                 <NA>
#> 3  Blattläuse (Röhrenläuse)             Aprikose
#> 31 Blattläuse (Röhrenläuse)              Kirsche
#> 32 Blattläuse (Röhrenläuse) Pfirsich / Nektarine
#> 33 Blattläuse (Röhrenläuse)              Pflaume
#> 34 Blattläuse (Röhrenläuse)            Zwetschge
#> 4               Spinnmilben        Obstbau allg.
# }