Create a dm object from an XML version of the PSMV

Create a dm object from an XML version of the PSMV

Usage

srppp_dm(from = srppp_xml_url, remove_duplicates = TRUE)

# S3 method for srppp_dm
print(x, ...)

Arguments

from

A specification of the way to retrieve the XML

remove_duplicates

Should duplicates based on wNbrs be removed?

x

A srppp_dm object

...

Not used

Value

A dm object with tables linked by foreign keys pointing to primary keys, i.e. with referential integrity.

Examples

# \dontrun{
library(dm)

srppp_cur <- srppp_dm()
dm_examine_constraints(srppp_cur)
#> ℹ All constraints satisfied.
dm_draw(srppp_cur)

# Show some information for products named 'Boxer'
srppp_cur |>
  dm_filter(products = (name == "Boxer")) |>
  dm_nrow()
#>             products                pNbrs   product_categories 
#>                    2                    1                    2 
#>    formulation_codes     parallel_imports       danger_symbols 
#>                    2                    8                    4 
#>                CodeS                CodeR         signal_words 
#>                    4                    8                    2 
#>           substances          ingredients                 uses 
#>                    1                    1                   34 
#> application_comments        culture_forms             cultures 
#>                   36                   18                   78 
#>                pests          obligations 
#>                   68                  230 
# }