Get Products from an XML version of the Swiss Register of Plant Protection Products

Get Products from an XML version of the Swiss Register of Plant Protection Products

Usage

srppp_xml_get_products(
  srppp_xml = srppp_xml_get(),
  verbose = TRUE,
  remove_duplicates = TRUE
)

Arguments

srppp_xml

An object as returned by 'srppp_xml_get'

verbose

Should we give some feedback?

remove_duplicates

Should duplicates based on wNbrs be removed? If set to 'TRUE', one of the two entries with identical wNbrs is removed, based on an investigation of background information carried out by the package authors. In all cases except for one, one of the product sections with duplicate wNbrs has information about an expiry of the registration, and the other doesn't. In these cases the registration without expiry is kept, and the expiring registration is discarded. In the remaining case (wNbr 5945), the second entry is selected, as it contains more indications which were apparently intended to be published as well.

Value

A tibble with a row for each product section in the XML file. An attribute 'duplicated_wNbrs' is also returned, containing duplicated W-Numbers, if applicable, or NULL.

Examples

# Get current list of products
srppp_xml_get_products()
#> # A tibble: 1,694 × 7
#>    wNbr   name          pNbr exhaustionDeadline soldoutDeadline isSalePermission
#>    <chr>  <chr>        <int> <chr>              <chr>           <chr>           
#>  1 1454   "Misto 12"    1865 ""                 ""              false           
#>  2 1454-2 "Parafol"     1865 ""                 ""              true            
#>  3 1526   "Zofal D"     1994 ""                 ""              false           
#>  4 1526-1 "Capito Win…  1994 ""                 ""              true            
#>  5 1526-2 "Biorga Con…  1994 ""                 ""              true            
#>  6 1529   "Oléoc "      2028 ""                 ""              false           
#>  7 1698   "Asulox"      2092 ""                 ""              false           
#>  8 18     "Thiovit Je…    38 ""                 ""              false           
#>  9 18-1   "Sufralo"       38 ""                 ""              true            
#> 10 18-2   "Capito Bio…    38 ""                 ""              true            
#> # ℹ 1,684 more rows
#> # ℹ 1 more variable: terminationReason <chr>