Create a dm object from an XML version of the Swiss Register of Plant Protection Products

In general, the information obtained from the XML file is left unchanged. For exceptions, see Details.

Usage

srppp_dm(from = srppp_xml_url, remove_duplicates = TRUE)

# S3 method for class 'srppp_dm'
print(x, ...)

Arguments

from

A specification of the way to retrieve the XML

remove_duplicates

Should duplicates based on wNbrs be removed?

x

A srppp_dm object

...

Not used

Value

A dm object with tables linked by foreign keys pointing to primary keys, i.e. with referential integrity.

Details

In the following case, the product composition is corrected while reading in the data:

The active substance content of Dormex (W-3066) is not 667 g/L, but 520 g/L This was confirmed by a visit to the Wädenswil archive by Johannes Ranke and Daniel Baumgartner, 2024-03-27.

In the table of obligations, the following information on mitigation measures is extracted from the ones relevant for the environment (SPe 3).

"sw_drift_dist": Unsprayed buffer towards surface waters to mitigate spray drift in meters

"sw_runoff_dist": Vegetated buffer towards surface waters to mitigate runoff in meters

"sw_runoff_points": Required runoff mitigation points to mitigate runoff

"biotope_drift_dist": Unsprayed buffer towards biotopes (as defined in articles 18a and 18b of the Federal Act on the Protection of Nature and Cultural Heritage) to mitigate spray drift in meters

Examples

library(dplyr, warn.conflicts = FALSE)
library(dm, warn.conflicts = FALSE)

sr <- srppp_dm()
dm_examine_constraints(sr)
#> ℹ All constraints satisfied.
# \dontrun{
dm_draw(sr)

# }

# Show ingredients for products named 'Boxer'
sr$products |>
  filter(name == "Boxer") |>
  left_join(sr$ingredients) |>
  left_join(sr$substances) |>
  select(wNbr, name, pNbr, isSalePermission, substance_de, g_per_L)
#> Joining with `by = join_by(pNbr)`
#> Joining with `by = join_by(pk)`
#> # A tibble: 2 × 6
#>   wNbr   name   pNbr isSalePermission substance_de g_per_L
#>   <chr>  <chr> <int> <chr>            <chr>          <dbl>
#> 1 6168   Boxer  7105 false            Prosulfocarb     800
#> 2 6168-1 Boxer  7105 true             Prosulfocarb     800

# Show authorised uses of the original product
boxer_uses <- sr$products |>
  filter(name == "Boxer", isSalePermission == "false") |>
  left_join(sr$uses) |>
  select(wNbr, name, use_nr,
    min_dosage, max_dosage, min_rate, max_rate, units_de,
    waiting_period, time_units_de, application_area_de)
#> Joining with `by = join_by(wNbr)`
print(boxer_uses)
#> # A tibble: 17 × 11
#>    wNbr  name  use_nr min_dosage max_dosage min_rate max_rate units_de
#>    <chr> <chr>  <int>      <dbl>      <dbl>    <dbl>    <dbl> <chr>   
#>  1 6168  Boxer      1         NA         NA      5       NA   l/ha    
#>  2 6168  Boxer      2         NA         NA      5       NA   l/ha    
#>  3 6168  Boxer      3         NA         NA      5       NA   l/ha    
#>  4 6168  Boxer      4         NA         NA      4       NA   l/ha    
#>  5 6168  Boxer      5         NA         NA      2.5      3   l/ha    
#>  6 6168  Boxer      6         NA         NA      4       NA   l/ha    
#>  7 6168  Boxer      7         NA         NA      3       NA   l/ha    
#>  8 6168  Boxer      8         NA         NA      3       NA   l/ha    
#>  9 6168  Boxer      9         NA         NA      4       NA   l/ha    
#> 10 6168  Boxer     10         NA         NA      5       NA   l/ha    
#> 11 6168  Boxer     11         NA         NA      4       NA   l/ha    
#> 12 6168  Boxer     12         NA         NA      3        4.5 l/ha    
#> 13 6168  Boxer     13         NA         NA      2.5      5   l/ha    
#> 14 6168  Boxer     14         NA         NA      3        5   l/ha    
#> 15 6168  Boxer     15         NA         NA      5       NA   l/ha    
#> 16 6168  Boxer     16         NA         NA      4       NA   l/ha    
#> 17 6168  Boxer     17         NA         NA      2.5      5   l/ha    
#> # ℹ 3 more variables: waiting_period <int>, time_units_de <chr>,
#> #   application_area_de <chr>

# Show crop for use number 1
boxer_uses |>
  filter(use_nr == 1) |>
  left_join(sr$cultures, join_by(wNbr, use_nr)) |>
  select(use_nr, culture_de)
#> # A tibble: 1 × 2
#>   use_nr culture_de     
#>    <int> <chr>          
#> 1      1 Knollensellerie

# Show target pests for use number 1
boxer_uses |>
  filter(use_nr == 1) |>
  left_join(sr$pests, join_by(wNbr, use_nr)) |>
  select(use_nr, pest_de)
#> # A tibble: 2 × 2
#>   use_nr pest_de                              
#>    <int> <chr>                                
#> 1      1 Einjährige Dicotyledonen (Unkräuter) 
#> 2      1 Einjährige Monocotyledonen (Ungräser)

# Show obligations for use number 1
boxer_uses |>
  filter(use_nr == 1) |>
  left_join(sr$obligations, join_by(wNbr, use_nr)) |>
  select(use_nr, sw_runoff_points, obligation_de) |>
  knitr::kable() |>
  print()
#> 
#> 
#> | use_nr|sw_runoff_points |obligation_de                                                                                                                                                                                                                                                                                                                                                                               |
#> |------:|:----------------|:-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------|
#> |      1|NA               |Nachbau anderer Kulturen: 16 Wochen Wartefrist.                                                                                                                                                                                                                                                                                                                                             |
#> |      1|NA               |Nachfolgearbeiten in behandelten Kulturen: bis 48 Stunden nach Ausbringung des Mittels Schutzhandschuhe + Schutzanzug tragen.                                                                                                                                                                                                                                                               |
#> |      1|NA               |Maximal 1 Behandlung pro Kultur.                                                                                                                                                                                                                                                                                                                                                            |
#> |      1|NA               |Ansetzen der Spritzbrühe: Schutzhandschuhe tragen. Ausbringen der Spritzbrühe: Schutzhandschuhe + Schutzanzug + Visier + Kopfbedeckung tragen. Technische Schutzvorrichtungen während des Ausbringens (z.B. geschlossene Traktorkabine) können die vorgeschriebene persönliche Schutzausrüstung ersetzen, wenn gewährleistet ist, dass sie einen vergleichbaren oder höheren Schutz bieten. |
#> |      1|1                |SPe 3: Zum Schutz von Gewässerorganismen muss das Abschwemmungsrisiko gemäss den Weisungen der Zulassungsstelle um 1 Punkt reduziert werden.                                                                                                                                                                                                                                                |
#> |      1|NA               |Splitbehandlung gemäss den Angaben der Bewilligungsinhaberin (max. 3 l/ha je Split, angegebene Aufwandmenge entspricht total bewilligter Menge).                                                                                                                                                                                                                                            |
#> |      1|NA               |Phytotoxschäden bei empfindlichen Arten oder Sorten möglich; vor allgemeiner Anwendung Versuchspritzung durchführen.                                                                                                                                                                                                                                                                        |
#> |      1|NA               |Bewilligt als geringfügige Verwendung nach Art. 35 PSMV (minor use).                                                                                                                                                                                                                                                                                                                        |

# Show application comments for use number 1
boxer_uses |>
  filter(use_nr == 1) |>
  left_join(sr$application_comments, join_by(wNbr, use_nr)) |>
  select(use_nr, application_comment_de)
#> # A tibble: 1 × 2
#>   use_nr application_comment_de      
#>    <int> <chr>                       
#> 1      1 "7 Tage nach dem Pflanzen. "