Get Parallel Imports from an XML version of the Swiss Register of Plant Protection Products

Get Parallel Imports from an XML version of the Swiss Register of Plant Protection Products

Usage

srppp_xml_get_parallel_imports(srppp_xml = srppp_xml_get())

Arguments

srppp_xml

An object as returned by 'srppp_xml_get'

Value

A tibble::tibble with a row for each parallel import section in the XML file.

Examples

# Get current list of parallel_imports
# \donttest{
srppp_xml_get_parallel_imports()
#> # A tibble: 746 × 9
#>    id     wNbr  name   admissionnumber producingCountryPrim…¹ exhaustionDeadline
#>    <chr>  <chr> <chr>  <chr>                            <int> <chr>             
#>  1 F-7056 1899  Life … 2159999                           6171 ""                
#>  2 I-2789 1899  Pirim… 4701                              6176 ""                
#>  3 B-5440 2054  Basam… 5675P/B                           6163 ""                
#>  4 F-3632 2054  Basam… 6800403                           6171 ""                
#>  5 I-6151 2054  Basam… 1573                              6176 ""                
#>  6 F-2010 2671  Fluid… 5100219                           6171 ""                
#>  7 F-7271 2935  Polyr… 8700107                           6171 "2025-07-01"      
#>  8 I-7270 2935  Polyr… 7916                              6176 "2025-07-01"      
#>  9 I-5565 3002  Dazid… 012455                            6176 ""                
#> 10 I-6469 3049  Bonzi  016305                            6176 ""                
#> # ℹ 736 more rows
#> # ℹ abbreviated name: ¹​producingCountryPrimaryKey
#> # ℹ 3 more variables: soldoutDeadline <chr>, pNbr <int>,
#> #   permission_holder <int>
# }